fixbug: comment contorller

+++ src/controllers/comment.controller.js
@@ -1,4 +1,4 @@
-import mongoose from "mongoose";
+import mongoose, { isValidObjectId } from "mongoose";
 import { Comment } from "../models/comment.model.js";
 import { ApiError } from "../utils/ApiError.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
@@ -8,6 +8,23 @@ const getVideoComments = asyncHandler(async (req, res) => {
     //TODO: get all comments for a video
     const { videoId } = req.params;
     const { page = 1, limit = 10 } = req.query;
+    const allVideoComments = await Comment.aggregate([
+        {
+            $match: {
+                video: new mongoose.Types.ObjectId(videoId)
+            }
+        }
+    ]);
+
+
+    Comment.aggregatePaginate(allVideoComments, { page, limit })
+        .then((result) => {
+            return res.status(201).json(
+                new ApiResponse(200, result, "VideoComments fetched  successfully!!"));
+        })
+        .catch((error) => {
+            throw new ApiError(500, "something went wrong while fetching video Comments", error);
+        });
 
 });
 
@@ -16,9 +33,14 @@ const addComment = asyncHandler(async (req, res) => {
     const { videoId } = req.params;
     const { comment } = req.body;
 
-    if (!comment) {
+    if (!isValidObjectId(videoId)) {
+        throw new ApiError(400, "This video id is not valid");
+    }
+
+    if (!comment || comment.trim() === "") {
         throw new ApiError(400, 'Please give a comment.');
     }
+
 
     const commentData = await Comment.create({
         video: videoId,
@@ -37,6 +59,21 @@ const updateComment = asyncHandler(async (req, res) => {
     // TODO: update a comment
     const { commentId } = req.params;
     const { comment } = req.body;
+
+    if (!isValidObjectId(commentId)) {
+        throw new ApiError(400, "This video id is not valid");
+    }
+
+    const findComment = await Comment.findById(commentId);
+
+    if (!findComment) {
+        throw new ApiError(404, "comment not found!");
+    }
+
+    if (findComment.owner.toString() !== req.user._id.toString()) {
+        throw new ApiError(403, "You don't have permission to update this comment!");
+    }
+
 
     const updatedComment = await Comment.findByIdAndUpdate(
         commentId,
@@ -58,6 +95,20 @@ const updateComment = asyncHandler(async (req, res) => {
 const deleteComment = asyncHandler(async (req, res) => {
     // TODO: delete a comment
     const { commentId } = req.params;
+
+    if (!isValidObjectId(commentId)) {
+        throw new ApiError(400, "This video id is not valid");
+    }
+
+    const findComment = await Comment.findById(commentId);
+
+    if (!findComment) {
+        throw new ApiError(404, "comment not found!");
+    }
+
+    if (findComment.owner.toString() !== req.user._id.toString()) {
+        throw new ApiError(403, "You don't have permission to update this comment!");
+    }
 
     const deletedComment = await Comment.findByIdAndDelete(commentId);
